Allow users to initiate Google Drive uploads directly from the left menu.
Allow users to initiate Google Drive uploads directly from the left menu.

+++ app/src/main/java/org/obd/graphs/preferences/trips/TripViewAdapter.kt
@@ -25,55 +25,55 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.CheckBox
+import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
-import org.obd.graphs.*
+import org.obd.graphs.R
 import org.obd.graphs.bl.trip.tripManager
 import org.obd.graphs.profile.profile
 import org.obd.graphs.ui.common.Colors
 import org.obd.graphs.ui.common.setText
 import java.text.SimpleDateFormat
-import java.util.*
+import java.util.Date
+import java.util.Locale
 
 private const val LOGGER_KEY = "TripsViewAdapter"
 
 class TripViewAdapter internal constructor(
     context: Context?,
-    var data: MutableCollection<TripFileDescDetails>
+    var data: MutableCollection<TripLogDetails>,
+    private val showDeleteButton: Boolean = true,
 ) : RecyclerView.Adapter<TripViewAdapter.ViewHolder>() {
-
     private val mInflater: LayoutInflater = LayoutInflater.from(context)
     private val dateFormat: SimpleDateFormat =
         SimpleDateFormat("yyyy.MM.dd HH:mm:ss", Locale.getDefault())
 
-    private val profileColors = mutableMapOf<String, Int>().apply {
-        val colors = Colors().get()
-        profile.getAvailableProfiles().forEach { (s, _) ->
-            if (colors.hasNext()) {
-                put(s, colors.nextInt())
+    private val profileColors =
+        mutableMapOf<String, Int>().apply {
+            val colors = Colors().get()
+            profile.getAvailableProfiles().forEach { (s, _) ->
+                if (colors.hasNext()) {
+                    put(s, colors.nextInt())
+                }
             }
         }
-    }
 
     override fun onCreateViewHolder(
         parent: ViewGroup,
-        viewType: Int
-    ): ViewHolder {
-        return ViewHolder(mInflater.inflate(R.layout.item_trip, parent, false))
-    }
+        viewType: Int,
+    ): ViewHolder = ViewHolder(mInflater.inflate(R.layout.item_trip, parent, false))
 
     override fun onBindViewHolder(
         holder: ViewHolder,
-        position: Int
+        position: Int,
     ) {
-
         data.elementAt(position).run {
             holder.vehicleProfile.setText(
                 source.profileLabel,
                 profileColors[source.profileId]!!,
                 Typeface.NORMAL,
-                0.6f
+                0.6f,
             )
             var startTs = source.startTime
             source.startTime.toLongOrNull()?.let {
@@ -103,44 +103,49 @@ class TripViewAdapter internal constructor(
         }
     }
 
-    override fun getItemCount(): Int {
-        return data.size
-    }
+    override fun getItemCount(): Int = data.size
 
-    inner class ViewHolder internal constructor(binding: View) :
-        RecyclerView.ViewHolder(binding) {
+    inner class ViewHolder internal constructor(
+        binding: View,
+    ) : RecyclerView.ViewHolder(binding) {
         val selected: CheckBox = binding.findViewById(R.id.trip_selected)
         val vehicleProfile: TextView = binding.findViewById(R.id.trip_profile)
         val tripStartDate: TextView = binding.findViewById(R.id.trip_start_date)
         val tripTime: TextView = binding.findViewById(R.id.trip_length)
         private val loadTrip: Button = binding.findViewById(R.id.trip_load)
         private val deleteTrip: Button = binding.findViewById(R.id.trip_delete)
+        private val actionsContainer: LinearLayout = binding.findViewById(R.id.actions_container)
 
         init {
+            if (showDeleteButton) {
+                loadTrip.setOnClickListener {
+                    tripManager.loadTripAsync(data.elementAt(adapterPosition).source.fileName)
+                }
 
-            loadTrip.setOnClickListener {
-                tripManager.loadTripAsync( data.elementAt(adapterPosition).source.fileName)
-            }
+                deleteTrip.setOnClickListener {
+                    val builder = AlertDialog.Builder(binding.context)
+                    val title = binding.context.getString(R.string.trip_delete_dialog_ask_question)
+                    val yes = binding.context.getString(R.string.trip_delete_dialog_ask_question_yes)
+                    val no = binding.context.getString(R.string.trip_delete_dialog_ask_question_no)
 
-            deleteTrip.setOnClickListener {
-                val builder = AlertDialog.Builder(binding.context)
-                val title = binding.context.getString(R.string.trip_delete_dialog_ask_question)
-                val yes = binding.context.getString(R.string.trip_delete_dialog_ask_question_yes)
-                val no = binding.context.getString(R.string.trip_delete_dialog_ask_question_no)
-
-                builder.setMessage(title)
-                    .setCancelable(false)
-                    .setPositiveButton(yes) { _, _ ->
-                        val trip = data.elementAt(adapterPosition)
-                        Log.i(LOGGER_KEY, "Trip selected to delete: $trip")
-                        data.remove(trip)
-                        tripManager.deleteTrip(trip.source)
-                        notifyDataSetChanged()
-                    }
-                    .setNegativeButton(no) { dialog, _ ->
-                        dialog.dismiss()
-                    }
+                    builder
+                        .setMessage(title)
+                        .setCancelable(false)
+                        .setPositiveButton(yes) { _, _ ->
+                            val trip = data.elementAt(adapterPosition)
+                            Log.i(LOGGER_KEY, "Trip selected to delete: $trip")
+                            data.remove(trip)
+                            tripManager.deleteTrip(trip.source)
+                            notifyDataSetChanged()
+                        }.setNegativeButton(no) { dialog, _ ->
+                            dialog.dismiss()
+                        }
                     builder.create().show()
+                }
+            } else {
+                loadTrip.visibility = View.GONE
+                deleteTrip.visibility = View.GONE
+                actionsContainer.visibility = View.GONE
             }
         }
     }
